+++ src/app/models/Country.ts
@@ -1,0 +1,10 @@
+export interface Country {
+  // Model for country data as structured in the package 'country-finder'
+  id: number;
+  name: string;
+  iso2: string;
+  iso3: string;
+  continent: string;
+  lat: string;
+  long: string;
+}
